Moved to a single contract
