+++ solana/energy-trading/programs/Energy-Market/src/errors.rs
@@ -1,4 +1,3 @@
-
 use anchor_lang::prelude::*;
 
 #[error_code]
