--- solana/energy-trading/programs/energy-injection/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "energy-injection"
-version = "0.1.0"
-description = "Created with Anchor"
-edition = "2021"
-
-[lib]
-crate-type = ["cdylib", "lib"]
-name = "energy_injection"
-
-[features]
-no-entrypoint = []
-no-idl = []
-no-log-ix-name = []
-cpi = ["no-entrypoint"]
-default = []
-
-[dependencies]
-anchor-lang = {version = "0.25.0"}
